+++ packages/article-topics/src/styles/index.js
@@ -1,13 +1,12 @@
-import { StyleSheet } from "react-native";
 import sharedStyles from "./shared";
 
-const styles = StyleSheet.create({
+const styles = {
   ...sharedStyles,
   text: {
     ...sharedStyles.text,
     // Gill Sans hack
     top: 2
   }
-});
+};
 
 export default styles;